requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyGithub==2.1.1
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
APScheduler==3.10.4
python-dateutil==2.8.2
zhipuai==1.0.7 